fix: address convention violations in handler base classes
- Export Span type from OpenTelemetry vendor wrapper
- Update handler imports to use vendor wrapper instead of @opentelemetry/api
- Move handler interfaces (SqsRecordContext, S3RecordContext, ScheduledResult)
  to src/types/lambda.ts per types-location convention
- Shorten @example tags to meet comment-conventions max 3 lines
