+++ src/lib/lambda/handlers/index.ts
@@ -4,22 +4,7 @@
  * Provides abstract base classes for all Lambda handler types with built-in observability.
  * Uses TypeScript decorators for declarative metrics, logging, and tracing.
  *
- * @example
- * ```typescript
- * import {AuthenticatedHandler, logger, metrics, MetricUnit} from '#lib/lambda/handlers'
- *
- * class MyHandler extends AuthenticatedHandler {
- *   readonly operationName = 'MyOperation'
- *
- *   protected async handleAuthenticated(event, context) {
- *     // Business logic here
- *     // this.userId is guaranteed to be a valid string
- *   }
- * }
- *
- * const handlerInstance = new MyHandler()
- * export const handler = handlerInstance.handler.bind(handlerInstance)
- * ```
+ * @example See any Lambda in src/lambdas/ for implementation examples
  */
 
 // Base handler with core decorators
@@ -32,18 +17,18 @@ export { OptionalAuthHandler } from './OptionalAuthHandler'
 
 // SQS handlers
 export { SqsHandler } from './SqsHandler'
-export type { SqsBatchOptions, SqsRecordContext } from './SqsHandler'
 
 // S3 event handlers
 export { S3EventHandler } from './S3EventHandler'
-export type { S3RecordContext } from './S3EventHandler'
 
 // Scheduled handlers
 export { ScheduledHandler } from './ScheduledHandler'
-export type { ScheduledResult } from './ScheduledHandler'
 
 // Invoke handlers (manual/Lambda-to-Lambda)
 export { InvokeHandler } from './InvokeHandler'
 
 // Custom authorizer handlers
 export { AuthorizerHandler } from './AuthorizerHandler'
+
+// Re-export types from centralized types module
+export type { S3RecordContext, ScheduledResult, SqsBatchOptions, SqsRecordContext } from '#types/lambda'
